Moved export config button to header in standalone
