+++ src/components/side-menu.vue
@@ -283,6 +283,7 @@
         </ul>
 
         <div v-show="sidemenuStore.expanded">
+            <!-- Import button -->
             <button
                 class="flex bg-black text-white rounded justify-center border border-black w-full hover:bg-gray-900 font-bold px-4 py-2 my-2"
                 tabindex="0"
@@ -315,68 +316,20 @@
                 :aria-label="$t('HACK.toc.importChart')"
                 @change="handleConfigFileUpload"
             />
-            <button
-                class="flex bg-white justify-center rounded border border-black w-full hover:bg-gray-100 font-bold px-4 py-2 my-2"
-                :aria-label="$t('HACK.toc.exportConfig')"
-                :class="{ 'disabled hover:bg-gray-400': btnDisabled }"
-                :disabled="btnDisabled"
-                :tabIndex="btnDisabled ? -1 : 0"
-                @click="exportHighchartsConfig"
-            >
-                <svg
-                    class="flex-shrink-0"
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="24"
-                    height="24"
-                    stroke="#000000"
-                    viewBox="0 0 24 24"
-                >
-                    <path d="M6 21H18M12 3V17M12 17L17 12M12 17L7 12" stroke-width="2" stroke-linecap="round"></path>
-                </svg>
-                {{ $t('HACK.toc.exportConfig') }}
-            </button>
-
-            <a
-                class="absolute right-2 bottom-2"
-                href="https://www.youtube.com/watch?v=zzk0VQ0dVMU&feature=youtu.be"
-                target="_blank"
-                v-if="chartStore.chartConfig && chartStore.chartConfig.title?.text?.en.toLowerCase() === 'breakfast'"
-            >
-                <svg width="24" height="24" viewBox="0 0 100 100" xmlns="http://www.w3.org/2000/svg" fill="#000000">
-                    <g id="SVGRepo_iconCarrier">
-                        <path
-                            fill="#F3F3F3"
-                            d="M67.283 96.847c-31.675 10.868-26.284-8.737-53.132-23.014-26.776-14.24-12.708-42.414 18.464-60.401 37.712-21.76 74.573-17.987 66.18 17.987-8.393 35.972-4.946 55.519-31.512 65.428z"
-                        ></path>
-                        <circle fill="#F0C419" cx="50" cy="47" r="19"></circle>
-                        <path
-                            stroke="#ffffff"
-                            stroke-width="4"
-                            stroke-linecap="round"
-                            stroke-linejoin="round"
-                            stroke-miterlimit="10"
-                            d="M79 7.031c6.138 0 13.807 6.252 15.077 14.969M53 35.378c3.757 1.081 6.71 3.965 8.025 7.622"
-                            fill="none"
-                        ></path>
-                    </g>
-                </svg>
-            </a>
+            <!-- Export/download button -->
+            <ChartExport :in-header="false" />
         </div>
     </nav>
 </template>
 
 <script setup lang="ts">
 import { computed, ref } from 'vue';
-import { useI18n } from 'vue-i18n';
 import { useDataStore } from '../stores/dataStore';
 import { useChartStore } from '../stores/chartStore';
 import { useSidemenuStore } from '../stores/sidemenuStore';
-import { saveAs } from 'file-saver';
 import { CurrentView } from '../definitions';
-import type { LangId } from '../definitions';
-import { buildContextMenuLabels } from '@/utils/contextMenu';
 
-import JSZip from 'jszip';
+import ChartExport from './helpers/chart-export.vue';
 
 const emit = defineEmits(['change-view']);
 
@@ -392,7 +345,6 @@ defineProps({
     }
 });
 
-const { t } = useI18n();
 const chartStore = useChartStore();
 const dataStore = useDataStore();
 const sidemenuStore = useSidemenuStore();
@@ -430,58 +382,6 @@ const handleConfigFileUpload = (event: Event) => {
     // clear input
     (event.target as HTMLInputElement).value = '';
     dataStore.setDatatableView(true);
-};
-
-const localizeConfig = (chartConfig: any, lang: string): any => {
-    if (Array.isArray(chartConfig)) {
-        return chartConfig.map((item) => localizeConfig(item, lang));
-    }
-
-    if (chartConfig && typeof chartConfig === 'object') {
-        // if it's a translation object
-        if ('en' in chartConfig || 'fr' in chartConfig) {
-            return chartConfig[lang] || '';
-        }
-
-        // otherwise recurse
-        const result = {};
-        for (const key in chartConfig) {
-            result[key] = localizeConfig(chartConfig[key], lang);
-        }
-        return result;
-    }
-
-    return chartConfig;
-};
-
-const exportHighchartsConfig = async () => {
-    // if bilingual charts is enabled, export both lang configs, otherwise just the uploaded/page file lang config
-    const langsToExport = chartStore.isBilingual ? ['en', 'fr'] : [chartStore.activeLang];
-    const files = langsToExport.map((lang) => {
-        // stringify + create blob for exported config
-        const localized = localizeConfig(chartStore.chartConfig, lang);
-
-        // override the context menu translations for this language
-        localized.lang = buildContextMenuLabels(t, lang as LangId);
-
-        const filename = `${localized.title?.text?.toLowerCase() || 'chart'}-${lang}.json`;
-        const json = JSON.stringify(localized, null, 2);
-
-        return { filename, json };
-    });
-
-    if (files.length === 1) {
-        // single file (no zip)
-        const blob = new Blob([files[0].json], { type: 'application/json' });
-        saveAs(blob, files[0].filename);
-    } else {
-        const zip = new JSZip();
-        files.forEach(({ filename, json }) => zip.file(filename, json));
-        const zipName = `${chartStore.chartConfig.title?.text[dataStore.uploadedFileLang].toLowerCase()}.zip`;
-
-        const content = await zip.generateAsync({ type: 'blob' });
-        saveAs(content, zipName);
-    }
 };
 </script>
